+++ fast/app/routers/system.py
@@ -5,4 +5,4 @@ router = APIRouter()
 
 @router.get("/healthcheck")
 async def healthcheck():
-    return Response(status_code=status.HTTP_200_OK)
+    return {"status": "ok"}
