+++ react-app/Dockerfile
@@ -1,13 +1,21 @@
-# ビルド環境
-FROM node:16 AS build
-WORKDIR /react-app
-COPY package.json package-lock.json ./
-RUN npm install
-COPY . .
-RUN npm run build
+# ベースとなるイメージを指定
+FROM node:14
 
-# 本番環境
-FROM nginx:alpine
-COPY --from=build /app/build /usr/share/nginx/html
-EXPOSE 80
-CMD ["nginx", "-g", "daemon off;"]
+# 作業ディレクトリを設定
+WORKDIR /app
+
+# 依存関係のファイルをコピー
+COPY package*.json ./
+
+# 依存関係をインストール
+RUN npm install
+RUN npm install --save-dev typescript @types/react @types/react-native @types/jest
+
+# アプリケーションのソースをコピー
+COPY . .
+
+# アプリケーションが起動するポートを指定
+EXPOSE 3000
+
+# アプリケーションを起動するコマンドを指定
+CMD ["npm", "start"]
